requirements.txt
Flask
Flask-SQLAlchemy
pandas
pyjwt
werkzeug